pyproject.toml
[tool.poetry]
name = "neuro_net"
version = "0.1.0"
description = ""
authors = ["Noriaki Mawatari <noriakimawatari@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-cov = "^2.12.1"
pre-commit = "^2.13.0"
flake8 = "^3.9.2"
mypy = "^0.910"
isort = "^5.9.2"
black = {version = "^21.7b0", allow-prereleases = true}
Flask = "^2.0.1"
geocoder = "^1.38.1"
yandex-geocoder = "^2.0.0"
python-decouple = "^3.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 79

[tool.black]
line-length = 79
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
